Remove some JITHelpers methods from Java 17+
This commit removes some JITHelpers methods from Java 17+, as they are
used only by the String implementation before Java 17.

- intrinsicIndexOfStringLatin1()
- intrinsicIndexOfStringUTF16()
- intrinsicIndexOfLatin1()
- intrinsicIndexOfUTF16()
