+++ jcl/src/java.base/share/classes/com/ibm/jit/JITHelpers.java
@@ -578,7 +578,6 @@ public final class JITHelpers {
 		}
 		return true;
 	}
-/*[ENDIF] JAVA_SPEC_VERSION < 17 */
 
 	/**
 	 * Returns the first index of the target character array within the source character array starting from the specified
@@ -723,6 +722,7 @@ public final class JITHelpers {
 		}
 		return -1;
 	}
+/*[ENDIF] JAVA_SPEC_VERSION < 17 */
 
 	/*
 	 * Constants for optimizedClone
